+++ src/route/state.rs
@@ -153,7 +153,7 @@ impl RouterState {
             "taguru.shard_call",
             otel.kind = "client",
             otel.name = %format!("{method} -> shard {shard}"),
-            taguru.shard.index = shard,
+            taguru.shard.index = shard as i64,
             http.request.method = %method,
             http.response.status_code = tracing::field::Empty,
             taguru.shard.outcome = tracing::field::Empty,
